requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
ultralytics==8.0.196
sahi==0.11.14
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
requests==2.31.0
gunicorn==21.2.0